+++ .env_exemple
@@ -7,15 +7,26 @@ DB_PORT=
 DB_HOST=
 DB_PASSWORD=
 
+POSTGRES_DB=
+POSTGRES_USER=
+POSTGRES_PASSWORD=
+
+CELERY_BROKER_URL=
+CELERY_RESULT_BACKEND=
+
+
+
 EMAIL_HOST_USER=
 EMAIL_PORT=
 EMAIL_HOST=
 EMAIL_PASSWORD=
 EMAIL_USE_TLS=
+EMAIL_USE_SSL=
+
+ACTIVATE_CODE_URL=
 
 LANGUAGE_CODE=
 
 TZ=
 
-DEBUG=True
-ACTIVATE_CODE_URL=http://localhost:8000/
+DEBUG=
