--- requirements.txt
@@ -1,50 +1,0 @@
-amqp
-asgiref
-async-timeout
-billiard
-celery
-certifi
-charset-normalizer
-click
-click-didyoumean
-click-plugins
-click-repl
-coreapi
-coreschema
-Deprecated
-Django
-django-filter
-djangorestframework
-djangorestframework-simplejwt
-drf-yasg
-idna
-inflection
-itypes
-Jinja2
-kombu
-MarkupSafe
-packaging
-Pillow
-prompt-toolkit
-psycopg2-binary
-PyJWT
-pyparsing
-python-decouple
-python-slugify
-pytz
-redis
-requests
-ruamel.yaml
-ruamel.yaml.clib
-six
-sqlparse
-text-unidecode
-uritemplate
-urllib3
-vine
-wcwidth
-wrapt
-pyTelegramBotApi
-gunicorn
-slugify
-django-cors-headers
