+++ server/src/apps/books/admin.py
@@ -1,5 +1,9 @@
 from django.contrib import admin
-from .models import Books, BookCover, Genre
+from .models import Books, BookCover, Genre, BookState, Rating
+
+# admin.site.register(BookState)
+admin.site.register(Rating)
+
 class TabularInlineImage(admin.TabularInline):
     model = BookCover
     extra = 0
