+++ server/src/apps/books/apps.py
@@ -4,3 +4,6 @@ from django.apps import AppConfig
 class BooksConfig(AppConfig):
     default_auto_field = 'django.db.models.BigAutoField'
     name = 'apps.books'
+
+    def ready(self):
+        from . import signals
