+++ server/src/apps/books/signals.py
@@ -1,0 +1,8 @@
+# from django.db.models.signals import pre_save
+# from django.dispatch import receiver
+# from myapp.models import MyModel
+#
+#
+# @receiver(pre_save, sender=MyModel)
+# def my_handler(sender, **kwargs):
+#     ...
