+++ server/src/services/drf_yasg.py
@@ -1,10 +1,10 @@
+
 from django.contrib import admin
 from django.urls import path, include
-from django.conf.urls.static import static
 from rest_framework import permissions
 from drf_yasg.views import get_schema_view
 from drf_yasg import openapi
-from django.conf import settings
+
 
 schema_view = get_schema_view(
    openapi.Info(
@@ -20,11 +20,5 @@ schema_view = get_schema_view(
 )
 
 urlpatterns = [
-   path('', schema_view.with_ui('swagger', cache_timeout=0), name='schema-swagger-ui'),
-   path('admin/', admin.site.urls),
-   path('account/', include('apps.account.urls')),
-   path('library/', include('apps.books.urls')),
+    path('', schema_view.with_ui('swagger', cache_timeout=0), name='schema-swagger-ui'),
 ]
-
-if settings.DEBUG:
-    urlpatterns += static(settings.MEDIA_URL, document_root=settings.MEDIA_ROOT)
