--- ПРОЧТИ МЕНЯ ТОЖЕ.md
@@ -1,2 +1,0 @@
-# Library
-Poetry добавил он в папек Library там пропишите poetry shell и межете работать
