--- Прочти меня.md
@@ -1,3 +1,0 @@
-# Library
-чтобы запустить проект вам нужно создать .env файл с зависимостями а все нужные зависимости в .env_exemple написаны 
-Будут вопросы пишите
